requirements.txt
pip

flask
pyyaml
